packages/consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
database = {path = "../database"}
kafka-python-ng = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
